+++ Comet-API/src/main/java/com/cometproject/api/modules/BaseModule.java
@@ -5,7 +5,7 @@ import com.cometproject.api.config.ModuleConfig;
 import com.cometproject.api.events.Event;
 import com.cometproject.api.events.EventListenerContainer;
 import com.cometproject.api.game.GameContext;
-import com.cometproject.api.networking.messages.MessageEventHandler;
+import com.cometproject.api.networking.messages.IMessageEventHandler;
 import com.cometproject.api.networking.sessions.ISession;
 import com.cometproject.api.server.IGameService;
 
@@ -47,7 +47,7 @@ public abstract class BaseModule implements EventListenerContainer {
         this.getGameService().getEventHandler().registerEvent(event);
     }
 
-    public void registerMessage(MessageEventHandler messageEventHandler) {
+    public void registerMessage(IMessageEventHandler messageEventHandler) {
 
     }