+++ Comet-API/src/main/java/com/cometproject/api/networking/NetworkingContext.java
@@ -1,4 +1,0 @@
-package com.cometproject.api.networking;
-
-public class NetworkingContext {
-}