+++ Comet-Networking-API/src/main/java/com/cometproject/networking/api/INetworkingServer.java
@@ -1,0 +1,16 @@
+package com.cometproject.networking.api;
+
+import com.cometproject.networking.api.config.NetworkingServerConfig;
+import com.cometproject.api.networking.sessions.ISessionService;
+import com.cometproject.networking.api.messages.IMessageHandler;
+
+public interface INetworkingServer {
+
+    void start();
+
+    NetworkingServerConfig getServerConfig();
+
+    IMessageHandler getMessageHandler();
+
+    ISessionService getSessionService();
+}